requirements.txt
opencv-python~=4.11.0.86
paddlepaddle~=3.0.0
paddleocr~=3.0.2
tqdm~=4.67.1
ultralytics~=8.3.156
numpy~=2.3.0
PyYAML~=6.0.2
matplotlib~=3.10.3
pillow~=11.2.1
pandas~=2.3.0